Fix #1531: Revert control key remapping from grave/tilde to zenkaku
The bug was introduced in the pull request #1214, and even in the description of the request it had been acknowledged that there is no such key on mac keyboards.

As a consequence, it led to inability to transfer the grave accent and and the tilde, at least from macOS to Windows on default "ABC" keyboard layout.

Unfortunately, I don't have a JIS keyboard, nor understand how to use it. The only thing I can suggest now is to revert the controversial remapping.